\name{RHawkes-package}
\alias{RHawkes-package}
\alias{RHawkes}
\docType{package}
\title{
\packageTitle{RHawkes}
}
\description{
\packageDescription{RHawkes}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RHawkes}
\packageIndices{RHawkes}

}
\author{
\packageAuthor{RHawkes}

Maintainer: \packageMaintainer{RHawkes}
}

\keyword{ package }
\keyword{ Hawkes process }
